+++ src/container/container.ts
@@ -14,7 +14,7 @@ import { ContainerSnapshot } from "./container_snapshot";
 import { Lookup } from "./lookup";
 import { ModuleActivationStore } from "./module_activation_store";
 
-type GetArgs = Omit<interfaces.NextArgs,'contextInterceptor'|'targetType'>
+type GetArgs<T> = Omit<interfaces.NextArgs<T>,'contextInterceptor'|'targetType'>
 
 class Container implements interfaces.Container {
 
@@ -27,7 +27,6 @@ class Container implements interfaces.Container {
     private _deactivations: interfaces.Lookup<interfaces.BindingDeactivation<any>>;
     private _snapshots: interfaces.ContainerSnapshot[];
     private _metadataReader: interfaces.MetadataReader;
-    private _appliedMiddleware: interfaces.Middleware[] = [];
     private _moduleActivationStore: interfaces.ModuleActivationStore
 
     public static merge(
@@ -250,6 +249,11 @@ class Container implements interfaces.Container {
         return bound;
     }
 
+    // check binding dependency only in current container
+    public isCurrentBound<T>(serviceIdentifier: interfaces.ServiceIdentifier<T>): boolean {
+        return this._bindingDictionary.hasKey(serviceIdentifier);
+    }
+
     public isBoundNamed(serviceIdentifier: interfaces.ServiceIdentifier<any>, named: string | number | symbol): boolean {
         return this.isBoundTagged(serviceIdentifier, METADATA_KEY.NAMED_TAG, named);
     }
@@ -302,7 +306,6 @@ class Container implements interfaces.Container {
     }
 
     public applyMiddleware(...middlewares: interfaces.Middleware[]): void {
-        this._appliedMiddleware = this._appliedMiddleware.concat(middlewares);
         const initial: interfaces.Next = (this._middleware) ? this._middleware : this._planAndResolve();
         this._middleware = middlewares.reduce(
             (prev, curr) => curr(prev),
@@ -390,13 +393,15 @@ class Container implements interfaces.Container {
     }
 
     public resolve<T>(constructorFunction: interfaces.Newable<T>) {
-        const tempContainer = this.createChild();
-        tempContainer.bind<T>(constructorFunction).toSelf();
-        this._appliedMiddleware.forEach((m) => {
-            tempContainer.applyMiddleware(m);
-        });
-
-        return tempContainer.get<T>(constructorFunction);
+        const isBound = this.isBound(constructorFunction);
+        if (!isBound) {
+            this.bind<T>(constructorFunction).toSelf();
+        }
+		const resolved =  this.get<T>(constructorFunction);
+        if (!isBound) {
+            this.unbind(constructorFunction);
+        }
+        return resolved;
     }
 
     private _preDestroy(constructor: any, instance: any): Promise<void> | void {
@@ -545,14 +550,14 @@ class Container implements interfaces.Container {
         });
 
     }
-    private _getAll<T>(getArgs:GetArgs): Promise<T[]>{
+    private _getAll<T>(getArgs:GetArgs<T>): Promise<T[]>{
         return Promise.all(this._get<T>(getArgs) as (Promise<T>|T)[]);
     }
     // Prepares arguments required for resolution and
     // delegates resolution to _middleware if available
     // otherwise it delegates resolution to _planAndResolve
-    private _get<T>(getArgs: GetArgs): interfaces.ContainerResolution<T> {
-        const planAndResolveArgs:interfaces.NextArgs = {
+    private _get<T>(getArgs: GetArgs<T>): interfaces.ContainerResolution<T> {
+        const planAndResolveArgs:interfaces.NextArgs<T> = {
             ...getArgs,
             contextInterceptor:(context) => context,
             targetType: TargetTypeEnum.Variable
@@ -569,7 +574,7 @@ class Container implements interfaces.Container {
     }
 
     private _getButThrowIfAsync<T>(
-        getArgs: GetArgs,
+        getArgs: GetArgs<T>,
     ): (T | T[]) {
         const result = this._get<T>(getArgs);
 
@@ -580,8 +585,8 @@ class Container implements interfaces.Container {
         return result as (T | T[]);
     }
 
-    private _getAllArgs<T>(serviceIdentifier: interfaces.ServiceIdentifier<T>): GetArgs {
-        const getAllArgs: GetArgs = {
+    private _getAllArgs<T>(serviceIdentifier: interfaces.ServiceIdentifier<T>): GetArgs<T> {
+        const getAllArgs: GetArgs<T> = {
             avoidConstraints: true,
             isMultiInject: true,
             serviceIdentifier,
@@ -595,8 +600,8 @@ class Container implements interfaces.Container {
         isMultiInject: boolean,
         key?: string | number | symbol,
         value?: any,
-    ): GetArgs {
-        const getNotAllArgs: GetArgs = {
+    ): GetArgs<T> {
+        const getNotAllArgs: GetArgs<T> = {
             avoidConstraints: false,
             isMultiInject,
             serviceIdentifier,
@@ -610,8 +615,8 @@ class Container implements interfaces.Container {
     // Planner creates a plan and Resolver resolves a plan
     // one of the jobs of the Container is to links the Planner
     // with the Resolver and that is what this function is about
-    private _planAndResolve<T>(): (args: interfaces.NextArgs) => interfaces.ContainerResolution<T> {
-        return (args: interfaces.NextArgs) => {
+    private _planAndResolve<T>(): (args: interfaces.NextArgs<T>) => interfaces.ContainerResolution<T> {
+        return (args: interfaces.NextArgs<T>) => {
 
             // create a plan
             let context = plan(
@@ -637,7 +642,7 @@ class Container implements interfaces.Container {
     }
 
     private _deactivateIfSingleton(binding: Binding<any>): Promise<void> | void {
-        if (!binding.cache) {
+        if (!binding.activated) {
             return;
         }