+++ test/annotation/inject.test.ts
@@ -3,10 +3,12 @@ declare function __param(paramIndex: number, decorator: ParameterDecorator): Cla
 
 import { expect } from "chai";
 import { decorate } from "../../src/annotation/decorator_utils";
-import { inject, LazyServiceIdentifer } from "../../src/annotation/inject";
+import { inject } from "../../src/annotation/inject";
+import { LazyServiceIdentifer } from "../../src/annotation/lazy_service_identifier";
 import * as ERROR_MSGS from "../../src/constants/error_msgs";
 import * as METADATA_KEY from "../../src/constants/metadata_keys";
 import { interfaces } from "../../src/interfaces/interfaces";
+import { multiInject } from "../../src/inversify";
 
 interface Katana {}
 interface Shuriken {}
@@ -113,7 +115,7 @@ describe("@inject", () => {
 
   });
 
-  it("Should throw when not applayed to a constructor", () => {
+  it("Should throw when not applied to a constructor", () => {
 
     const useDecoratorOnMethodThatIsNotAConstructor = function() {
       __decorate([ __param(0, inject("Katana")) ],
@@ -173,6 +175,25 @@ describe("@inject", () => {
     // no more metadata should be available
     expect(paramsMetadata["2"]).to.eq(undefined);
 
+  });
+
+  it("should throw when applied inject decorator with undefined service identifier to a property", () => {
+    expect(() => {
+      //@ts-ignore
+      class WithUndefinedInject{
+        @inject(undefined as any)
+        property:string
+      }
+    }).to.throw(`${ERROR_MSGS.UNDEFINED_INJECT_ANNOTATION("WithUndefinedInject")}`)
+  });
+
+  it("should throw when applied multiInject decorator with undefined service identifier to a constructor parameter", () => {
+    expect(() => {
+      //@ts-ignore
+      class WithUndefinedInject{
+        constructor(@multiInject(undefined as any) readonly dependency:string[]){}
+      }
+    }).to.throw(`${ERROR_MSGS.UNDEFINED_INJECT_ANNOTATION("WithUndefinedInject")}`)
   });
 
 });