+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- createTaggedDecorator #1343
 - Async bindings #1132
 - Async binding resolution (getAllAsync, getAllNamedAsync, getAllTaggedAsync, getAsync, getNamedAsync, getTaggedAsync, rebindAsync, unbindAsync, unbindAllAsync, unloadAsync) #1132
 - Global onActivation / onDeactivation #1132
@@ -15,6 +16,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - @postConstruct can target an asyncronous function #1132
+- Singleton scoped services cache resolved values once the result promise is fulfilled #1320
+
+### Fixed
+- only inject decorator can be applied to setters #1342
+- Container.resolve should resolve in that container #1338
+
+## [Unreleased]
+-Feat add API method for check dependency only in current container
 
 ## [5.1.1] - 2021-04-25
 -Fix pre-publish for build artifacts